Rename README.md Task 1: PROJECT APPROACH AND PROTOTYPE to README.md
User Experience Design explores multiple tools and techniques used in user experience design. 
Students are presented with an in-depth view of activities involved in the design of user experience 
and have the opportunity to create several deliverables, including persona profiles, information 
architectures, and prototypes of different levels of fidelity. In addition, the course also covers 
usability testing and the evaluation of quantitative and qualitative data derived from these and 
other experiments.